HW01 Optional major fix

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -3,82 +3,83 @@ package ru.javawebinar.topjava.web;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.javawebinar.topjava.model.Meal;
-import ru.javawebinar.topjava.model.MealTo;
 import ru.javawebinar.topjava.service.MealService;
 import ru.javawebinar.topjava.service.MealServiceImpl;
-import ru.javawebinar.topjava.util.MealsUtil;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.format.DateTimeParseException;
-import java.util.List;
-import java.util.Map;
-import java.util.function.BiFunction;
-import java.util.stream.Collectors;
+
 
 public class MealServlet extends HttpServlet {
     private static final Logger log = LoggerFactory.getLogger(MealServlet.class);
 
-    private final MealService service;
+    private MealService service;
 
-    public MealServlet() {
-        super();
-        this.service = new MealServiceImpl();
+    @Override
+    public void init() throws ServletException {
+        super.init();
+        service = new MealServiceImpl();
     }
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         log.debug("Redirect to meals");
-        String action = req.getParameter("action").toLowerCase();
-        switch (action) {
-            case "delete" : {
-                int mealId = Integer.parseInt(req.getParameter("mealID"));
-                service.delete(mealId);
-            }
-            break;
-            case "update" : {
-                int mealId = Integer.parseInt(req.getParameter("mealID"));
-                req.setAttribute("meal", service.getById(mealId));
-                req.getRequestDispatcher("/mealCreation.jsp").forward(req, resp);
-            }
-            break;
-            case "create" : {
-                req.getRequestDispatcher("/mealCreation.jsp").forward(req, resp);
+        String action = req.getParameter("action");
+        if (action != null) {
+            switch (action.toLowerCase()) {
+                case "delete": {
+                    log.debug("delete action");
+                    int mealId = service.parseIdFromRequest(req);
+                    service.delete(mealId);
+                    resp.sendRedirect("meals");
+                }
+                return;
+                case "update": {
+                    log.debug("update action");
+                    int mealId = service.parseIdFromRequest(req);
+                    req.setAttribute("meal", service.getById(mealId));
+                    req.getRequestDispatcher("/mealCreation.jsp").forward(req, resp);
+                }
+                break;
+                case "create": {
+                    log.debug("create action");
+                    req.getRequestDispatcher("/mealCreation.jsp").forward(req, resp);
+                }
             }
         }
 
-        req.setAttribute("meals", MealsUtil.filteredByStreams(MealsUtil.getMeals(), LocalTime.MIN, LocalTime.MAX, MealsUtil.CALORIES_LIMIT));
+        req.setAttribute("meals", service.getViewList());
         req.getRequestDispatcher("/meals.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("UTF-8");
-        LocalDateTime dateTime = null;
+        LocalDateTime dateTime;
         try {
             String dateTimeStr = req.getParameter("dateTime");
             dateTime = LocalDateTime.parse(dateTimeStr);
         } catch (DateTimeParseException e) {
             log.debug("DateTimeParseException appeared", e);
+            throw new ServletException("Exception during date parse");
         }
 
         String desc = req.getParameter("desc");
         int calories = Integer.parseInt(req.getParameter("cal"));
         String idStr = req.getParameter("id");
+        Meal newMeal = new Meal(dateTime, desc, calories);
         if (idStr == null || idStr.isEmpty()) {
-            service.add(new Meal(dateTime, desc, calories));
+            service.create(newMeal);
         } else {
-            Meal mealToBeUpdated = new Meal(dateTime, desc, calories);
-            mealToBeUpdated.setId(Integer.parseInt(idStr));
-            service.update(mealToBeUpdated);
+            newMeal.setId(Integer.parseInt(idStr));
+            service.update(newMeal);
         }
 
-        doGet(req, resp);
+       resp.sendRedirect("meals");
     }
 }
